8/31/25 rebase fixes

+++ include/graphics.h
@@ -1811,7 +1811,7 @@ extern const u32 gIntroFlygon_Gfx[];
 extern const u32 gPartyMenuBg_Gfx[];
 extern const u32 gPartyMenuBg_Tilemap[];
 extern const u16 gPartyMenuBg_Pal[];
-extern const u16 gPartyMenuBgYb_Gfx[];
+extern const u32 gPartyMenuBgYb_Gfx[];
 extern const u16 gPartyMenuBgYb_Pal[];
 
 // berry pics
